refactor(frontend): extract shared CodeSource interface (#2238)
Move the CodeSource interface — duplicated identically across 7 analytics
components and composables — into a single shared types/analytics.ts file.
All consumers now import from @/types/analytics.
