+++ autobot-frontend/src/types/analytics.ts
@@ -1,0 +1,25 @@
+// AutoBot - AI-Powered Automation Platform
+// Copyright (c) 2025 mrveiss
+// Author: mrveiss
+
+/**
+ * Shared types for codebase analytics components.
+ * Issue #1133: CodeSource type — extracted from 7 duplications (#2238).
+ */
+
+export interface CodeSource {
+  id: string
+  name: string
+  source_type: 'github' | 'local'
+  repo: string | null
+  branch: string
+  credential_id: string | null
+  clone_path: string | null
+  last_synced: string | null
+  status: 'configured' | 'syncing' | 'ready' | 'error'
+  error_message: string | null
+  owner_id: string | null
+  access: 'private' | 'shared' | 'public'
+  shared_with: string[]
+  created_at: string
+}
